后宮改名寵幸

+++ WorldOfTheThreeKingdoms/GameGlobal/Parameters.cs
@@ -459,7 +459,7 @@ namespace GameGlobal
                 AIExecuteMaxUncreulty = int.Parse(nextSibling.Attributes.GetNamedItem("AIExecuteMaxUncreulty").Value);
                 AIExecutePersonIdealToleranceMultiply = float.Parse(nextSibling.Attributes.GetNamedItem("AIExecutePersonIdealToleranceMultiply").Value);
 
-                AIHougongArchitectureCountProbMultiply = int.Parse(nextSibling.Attributes.GetNamedItem("AIHougongArchitectureCountProbMultiply").Value);
+                AIHougongArchitectureCountProbMultiply = float.Parse(nextSibling.Attributes.GetNamedItem("AIHougongArchitectureCountProbMultiply").Value);
                 AIHougongArchitectureCountProbPower = float.Parse(nextSibling.Attributes.GetNamedItem("AIHougongArchitectureCountProbPower").Value);
 
                 FireStayProb = int.Parse(nextSibling.Attributes.GetNamedItem("FireStayProb").Value);
